bunch of changes
figured out dark/light mode. needed to kill off a lot of app.css code and learn about how to wisely wrap the themeprovider. reference: https://ui.docs.amplify.aws/react/theming

making a markdumb renderer which is fun. keeping it simple

changing views to cards for the most part to work well with the theme systems. also removing background and text color props. can change it somewhere globally instead of like that.

added a lot of carbon stuff as I want to make that the example instead for cscale, next to a text explainer of the json format. it can be dummy data generated with ai based on specs and ranges; I don't want to waste electricity on examples here