change rule for eslint to ignore variables prefixed with _, or equal to _
otherwise lots of little fixes the compiler found when building. but not until it was in the cloud, uh oh.

i changed a lot about how the useCallback/useMemo are set up from looking at stuff online but I don't know if there's a point where I'm just doing more and this is all too tiny to care