Refactor out backend auth setup, refactor in the concept of earnestly using symbols as typing
Involves casting to String manually, but I watched a video about deoptimization with the v8 compiler, and I don't want it to do work twice over if I can help it.
Also simplified the firststreet jsons' signatures to omit a wrapping "data", making things more consistent and simple.

I still want to remove the transformers altogether so that wasn't a good use of time, oh well.